Rebase and final? polishing
