+++ spring-kafka/src/main/java/org/springframework/kafka/support/KafkaHeaders.java
@@ -115,18 +115,21 @@ public abstract class KafkaHeaders {
 	/**
 	 * The header containing information to correlate requests/replies.
 	 * Type: byte[].
+	 * @since 2.1.3
 	 */
 	public static final String CORRELATION_ID = PREFIX + "correlationId";
 
 	/**
 	 * The header containing the default reply topic.
 	 * Type: byte[].
+	 * @since 2.1.3
 	 */
 	public static final String REPLY_TOPIC = PREFIX + "replyTopic";
 
 	/**
 	 * The header containing a partition number on which to send the reply.
 	 * Type: binary (int) in byte[].
+	 * @since 2.1.3
 	 */
 	public static final String REPLY_PARTITION = PREFIX + "replyPartition";
